(1) Correção da operação create do controller/PlanetController.java. (2) Correção na operação delete do PlanetService. (3) Ajustes no application.properties.

+++ src/main/java/br/com/starwars/service/PlanetService.java
@@ -121,6 +121,12 @@ public class PlanetService extends AbstractService<Planet> {
 	}
 	
 	public void deletePlanet(Integer id) {
+		boolean exists = planetRepository.existsById(id);
+		
+		if (!exists) {
+			throw new ResourceNotFoundException("Planet com ID=" + String.valueOf(id) + " nao encontrado.");
+		}
+		
 		deleteById(id);
 	}
 }
